Simplify the test
